move the anim state graph to the anim directory
